refactor: api endpoints moved to endpoints.py while allowing app.py to run the same way
